docs: split up the weather app into more files
